+++ base/interactiveutil.jl
@@ -314,13 +314,13 @@ function versioninfo(io::IO=STDOUT; verbose::Bool=false, packages::Bool=false)
     println(io, "  LLVM: libLLVM-",libllvm_version," (", Sys.JIT, ", ", Sys.cpu_name, ")")
 
     println(io, "Environment:")
-    for (k,v) in ENV
+    for (k,v) in pairs(ENV)
         if ismatch(r"JULIA", String(k))
             println(io, "  $(k) = $(v)")
         end
     end
     if verbose
-        for (k,v) in ENV
+        for (k,v) in pairs(ENV)
             if ismatch(r"PATH|FLAG|^TERM$|HOME", String(k))
                 println(io, "  $(k) = $(v)")
             end